validate.js renamed to FormValidator.js
